Added CPU ticks to data logging + minor modifications of airframe

+++ sw/airborne/modules/loggers/sd_logger_spi_direct.h
@@ -68,7 +68,7 @@ struct LogPacket {
   int32_t data_8;
   int32_t data_9;
   int32_t data_10;
-  int32_t data_11;
+  float data_11;
   float data_12;
 };
